requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
Werkzeug==2.3.0
PyJWT==2.8.0
